+++ src/main/java/com/example/showadvisor_fia/SortIndividuiByNGeneri.java
@@ -6,8 +6,8 @@ public class SortIndividuiByNGeneri implements Comparator<Individuo> {
 
     @Override
     public int compare(Individuo o1, Individuo o2) {
-        Integer d = 0-o1.getnShowGeneri();
-        Integer d2 = 0-o2.getnShowGeneri();
+        Double d = 0-o1.getnShowGeneri();
+        Double d2 = 0-o2.getnShowGeneri();
         return d.compareTo(d2);
     }
 }
